+++ src/Lib/Error/ImageFileError.php
@@ -1,0 +1,9 @@
+<?php
+namespace App\Lib\Error;
+
+class ImageFileError
+{
+    public function __construct(public string $message)
+    {
+    }
+}
